requirements.txt
asgiref==3.6.0
django-debug-toolbar==4.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
Pillow==9.4.0
python-decouple==3.8
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3

